requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==3.2.2
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cryptography==45.0.7
databases==0.9.0
docx2txt==0.9
ebooklib==0.20
ecdsa==0.19.1
exceptiongroup==1.3.0
faiss-cpu==1.9.0
fastapi==0.116.1
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.30.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
multidict==6.6.4
numpy==2.0.2
packaging==25.0
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pytesseract==0.3.13
pgvector==0.4.1
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pypdfium2==4.30.0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
# Phase 3a — Celery 调度框架（机器人 / Agent B / Agent C 异步任务用）
celery==5.4.0
redis==5.0.8
